requirements/dev.txt
-r requirements.txt
torch==1.7.1
torchvision==0.8.2
hydra-configen==0.9.0.dev7
black==20.8b1
coverage
flake8==3.8.4
flake8-copyright
isort==5.5.2
mypy
nox
packaging
pre-commit
pytest
pytest-snail
setuptools
towncrier
twine
yamllint
